restructures app
