+++ project_electron/orgs/accession/admin.py
@@ -2,8 +2,9 @@
 from __future__ import unicode_literals
 
 from django.contrib import admin
-from accession.models import *
+from orgs.accession.models import *
+
 
 @admin.register(Accession)
 class AccessionAdmin(admin.ModelAdmin):
-	pass
+    pass
