+++ project_electron/orgs/accession/templates/accession/main.html
